allow application of @symbolic decorators without circular imports
